This transaction is in two parts:
1. Generator interface changes for IDE's

I have added to the default constructors of the Wadl2Java tool so that it is possible to have the source and destinations files taken from virtual file systems and the like. This shouldn't affect day to day usage; but will make it easier to plug into JDeveloper and other IDE's.

2. The Generator now produces code for the Jersey Client

I have removed the contents of the org/jvnet/ws/wadl/util/ module and now the code generated makes use of the Jersey Client, currently 1.8 but it should work for other versions.

The format of the non-JAXB method has changed slightly to allow parameters of Class or GeneticType to specify the return type rather rather than the old fashioned DataSource classes.

A @Generated annotation in added to the root class to make it easier to regenerate at a future date. The current generated code is likely to evolve - this is just a first cut conversion.

I have added a bunch more test cases and beefed up the current tests so that it verifies the methods that are generated have the correct parameters.