This transaction was the first steps to extracting the AST code from the generator code, currently this code was all stuck in the Wadl2Java class and now instead is embedded in the WadlAstBuilder class that can be used outside of the Wadl2Java project. There is still some work to do to remove codeModel dependencies from the ast model; but for the most part it is usable.
This should allow clients to share the code that solves the "href" references in the document both
for internal and external references. This might in some future interaction become a standard API
for accessing the structure of a WADL document in a way that is useful for clients and client generation
in a way that the JAX-B model doesn't entirely solve.