This is a fix for Oracle bug 14622829 where the exceptions generated by the WADL API were not actually used on the code.
So no a switch statement is generated where the correct exceptions are thrown if there is a 
valid JAX-B type; but just a plain WebApplicationException with at least the status code
and the full Response object in the case of RS 2.0.

All the exception classes now extend WebApplicationException to make the generated code more 
consistent with the published API. For RS 2.0 the response object is used in the normal way
but for Jersey 1.x an extra getClientResponse accessors is added. This is less of a pleasing
API but if the client only makes use of the status code then it would be clean.

The final change is to grab the AbstractMultivaluedMap class from JAX-RS 2.0 as I made it much
easier to create an ordered hash set of status codes. Since this an another Oracle project I 
guess this should be find to temporarily transplant the code until we can move up to a production
version of Jersey 2.x.