Oracle Bug 14831111
The client was no longer throwing an exception when there was a server fault unless there
as a registered fault type, this was cause of the code to access the status code for the
fault state.

Now the code will throw a generic WebApplicationException with the stauts code, and the
response object in -RS 2.0. 

Also this broke the ability toe pass in a (Client)Response as the Class<T> parameters
so it was no longer possible to access the response data directly.