This transaction revisits a change I made yesterday which was causing problems for the supplied version command line tool. Basically we do need a straight dependency on the JAX-RS 2.0 UriBuilder and my previous change was causing some confusion.
The problem now is that the version of JAX-B we are shipping requires a different version of the API 
than that is shipped with JDK 6, so I have done the work so that the command line versions of the tool
correctly configure the endorsed directory. (It does this for both JDK 6 & 7 as it is likely future
versions will also have this problem so easier to keep with a consistent version) I have also changed
the dependency type for the jaxb-api to provided where relevant as it doesn't help to have this jar on
the classpath as it will always be overriden by the JDK.

Users of Ant and Maven will still need to set up the endorsed directory for this to work in JDK 6 but this
can be documented at least and the command line tools works out of the box.

I tried to start writing some JUnit tests so that we didn't break this again, but for some reason the 
unit tests in wadl-dist were not being execute, I have checked in a single test to just