This transaction centralizes a few bits of .xsl that have until now be in different modules:
1) upgrade.xsl has now been moved to wadl-xslt module so that is is easier to reuse
2) xsl_2009-02.xsl is a HTML document generator that has been moved here so that it 
is easier to use for the OEPE product along with a helper java function. Currently 
this .xsl requires the Oracle XDK to run but hopefully this is will be fixed in a 
later revision

I have added static functions on WadlXslUtils to access a stream for both of these
transforms. 

I removed obsolete code from this class, and cleaned out the related Unit test.